Add a `BNFCode` model
This represents all levels of the hierarchy in a single table, which I
think is more ergonomic than separate Chapter, Section etc. models.

Note that we customise the database table name to remove the app name
prefix. Given that all tables in this database belong to a single app,
it's redundant. And because we anticipate using raw SQL more often than
in a standard Django project it's helpful to have less verbose table
names.