Simplify SQL by using `UNION ALL BY NAME`
This means we can use the same command to read all our files and not
worry about the fact that they may have different column names. DuckDB
will use the union of all column names as the columns for the result and
files which don't have a given column will just be given NULL values
there.