Add guard against non-unique group mappings
By "non-unique" we mean mapping the same input row into multiple output
groups. The code is perfectly capable of supporting this, but we can't
currently see the use-case and so we want to make sure we don't
accidentally do this for some reason. The guard is easy enough to remove
if we discover we do want to support this.