Only consider direct relations in `with_practice_ids()`
The NHS organisational structure (perhaps unsurprisingly) does not form
a neat hierarchy. For example, there are cases where a PCN combines
practices from multiple ICBs. In this case, the PCN will "belong" to a
particular ICB, but some of practices will belong to some other ICB.

If we take the naive transitive closure of the parent-child relationship
then we end up saying that a practice belongs to multiple ICBs, which is
not the case. Instead we only want to consider relationships as directly
supplied to us in the ODS data.