Use `CursorCacheKeyWrapper` in test fixture
This isn't absolutely essential: each DuckDB cursor object is already
considered unique so we won't end up inadvertently hitting the cache in
tests if we leave things as they are. But it's better to use the same
wrapper class here so as to better match production and make our tests
more meaningful.